+++ pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+markers =
+    notebook: Test runs a notebook - slow
+filterwarnings =
+    ignore::DeprecationWarning
+junit_family=legacy
