+++ tests/test_process_tree_utils.py
@@ -11,6 +11,7 @@ from pathlib import Path
 import nbformat
 from nbconvert.preprocessors import ExecutePreprocessor, CellExecutionError
 import pandas as pd
+import pytest
 
 from ..msticpy.sectools import process_tree_utils as ptutil
 
@@ -143,6 +144,7 @@ _NB_FOLDER = "docs/notebooks"
 _NB_NAME = "ProcessTree.ipynb"
 
 
+@pytest.mark.notebook
 def test_process_tree_notebook():
     nb_path = Path(_NB_FOLDER).joinpath(_NB_NAME)
     abs_path = Path(_NB_FOLDER).absolute()
