+++ tests/test_timeline.py
@@ -7,6 +7,7 @@ import unittest
 from pathlib import Path
 import nbformat
 from nbconvert.preprocessors import ExecutePreprocessor, CellExecutionError
+import pytest
 
 _NB_FOLDER = "docs/notebooks"
 _NB_NAME = "EventTimeline.ipynb"
@@ -15,6 +16,7 @@ _NB_NAME = "EventTimeline.ipynb"
 class TestnTimeline(unittest.TestCase):
     """Unit test class."""
 
+    @pytest.mark.notebook
     def test_timeline_controls(self):
         nb_path = Path(_NB_FOLDER).joinpath(_NB_NAME)
         abs_path = Path(_NB_FOLDER).absolute()
